Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
google-genai==0.3.0
python-dotenv==1.0.0
pandas==2.1.4
python-multipart==0.0.6
